requirements.txt
# Core ML and Deep Learning
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.3.0

# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Data Processing
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
python-json-logger>=2.0.0

# Optional: For production deployment
# gunicorn>=20.1.0